# docs/swagger.yaml
definitions:
  apis.IssueTokenRequest:
    properties:
      audience:
        type: string
      old_jti:
        type: string
      scopes:
        items:
          type: string
        type: array
      service:
        type: string
    type: object
  apis.PowerTableDoc:
    properties:
      company:
        example: Acme Corp
        type: string
      createdAt:
        example: "2026-04-23T10:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      numberOfHours:
        example: 5
        type: integer
      price:
        example: 99.99
        type: number
      selectionMode:
        example: price-based
        type: string
      updatedAt:
        example: "2026-04-23T10:00:00Z"
        type: string
      userId:
        example: user-abc-123
        type: string
    type: object
  apis.RevokedTokenDoc:
    properties:
      expiresAt:
        example: "2026-04-24T10:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      jti:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      revokedAt:
        example: "2026-04-23T10:00:00Z"
        type: string
    type: object
  apis.TokenResponse:
    properties:
      audience:
        type: string
      expires_at:
        type: string
      jti:
        type: string
      scopes:
        items:
          type: string
        type: array
      service:
        type: string
      token:
        type: string
    type: object
info:
  contact: {}
paths:
  /admin/tokens:
    post:
      consumes:
      - application/json
      description: Issues a new HS256 JWT for a named service with a specific audience
        and scopes.
      parameters:
      - description: Admin API key
        in: header
        name: X-Admin-Key
        required: true
        type: string
      - description: Token request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apis.IssueTokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apis.TokenResponse'
        "400":
          description: service, audience, and scopes are required
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Issue a service token
      tags:
      - admin
  /admin/tokens/{jti}:
    delete:
      description: Permanently revokes the token identified by its JTI. The token
        cannot be used after this call even if it has not yet expired.
      parameters:
      - description: Admin API key
        in: header
        name: X-Admin-Key
        required: true
        type: string
      - description: JWT ID (jti claim) of the token to revoke
        in: path
        name: jti
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Token not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Revoke a service token
      tags:
      - admin
  /admin/tokens/revoked:
    get:
      description: Returns the full revocation list. Useful for auditing and verifying
        that a token has been properly revoked.
      parameters:
      - description: Admin API key
        in: header
        name: X-Admin-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/apis.RevokedTokenDoc'
            type: array
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List revoked tokens
      tags:
      - admin
  /admin/tokens/rotate:
    post:
      consumes:
      - application/json
      description: Revokes the token identified by old_jti and issues a new token
        with the provided service, audience, and scopes.
      parameters:
      - description: Admin API key
        in: header
        name: X-Admin-Key
        required: true
        type: string
      - description: Rotation request (old_jti is required)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apis.IssueTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apis.TokenResponse'
        "400":
          description: service, audience, scopes, and old_jti are required
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Rotate a service token
      tags:
      - admin
  /charging:
    get:
      description: Returns the charging status for a given userId.
      parameters:
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get charging status
      tags:
      - charging
    patch:
      consumes:
      - application/json
      description: Updates the charging status for a given userId.
      parameters:
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      - description: Charging status
        in: body
        name: body
        required: true
        schema:
          properties:
            charging:
              type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Set charging status
      tags:
      - charging
  /power-table:
    get:
      description: Returns all power table entries. Optionally filter by userId query
        parameter.
      parameters:
      - description: Filter by user ID
        in: query
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/apis.PowerTableDoc'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List power table entries
      tags:
      - power-table
    post:
      consumes:
      - application/json
      description: Upserts a power table entry by userId. Updates company, price,
        selectionMode, and numberOfHours on conflict.
      parameters:
      - description: Power table entry
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apis.PowerTableDoc'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apis.PowerTableDoc'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create or update a power table entry
      tags:
      - power-table
swagger: "2.0"